+++ src/core/dates_names/daily_names.rs
@@ -11,7 +11,6 @@ use crate::prelude::*;
 use crate::core::date_models::units_validated::{
     ValidatedDate, ValidatedYear, ValidatedDay, ValidatedMonth,
 };
-
 #[derive(Debug, PartialEq, Eq, Getters, CopyGetters)]
 pub struct DailyName {
     #[getset(get_copy = "pub")]